Test : [ LottoTest 테스트케이스 수정 ]
