+++ __tests__/ApplicationTest.js
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines-per-function */
 import { MissionUtils } from '@woowacourse/mission-utils';
 import App from '../src/App.js';
